requirements.txt
# GIS and Data Processing
geopandas>=0.12.0             # Latest as of 2025, Python 3.12 compatible
rasterio>=1.3.6               # Latest with Python 3.12 support
numpy>=1.24.0
pandas==2.2.2                 # Upgraded from 1.5.3 for Python 3.12 support
scikit-learn==1.4.2           # Latest as of 2025
scipy>=1.10.0                 # Compatible with numpy 1.26
pyproj>=3.5.0                 # Latest stable
gdal==3.8.4                   # Requires GDAL system package; 3.8.x works with Python 3.12
shapely>=2.0.1                # Latest with full pyproj/geopandas compatibility

# Machine Learning and LLM
torch==2.2.2                  # Latest stable PyTorch (check CUDA compatibility if needed)
transformers==4.40.2          # Latest stable
openai>=1.0.0                 # Matches current OpenAI API client
langchain==0.1.16             # Refactored versioning, compatible with latest modules

# Web Framework
fastapi>=0.95.0               # Current latest
uvicorn>=0.21.0               # Latest as of 2025
pydantic>=2.0.0               # v2 line is stable and recommended

# Database
psycopg2-binary==2.9.9        # Maintained binary wheel
sqlalchemy>=2.0.0             # Up-to-date and stable
geoalchemy2>=0.14.4           # For spatial data types and operations
alembic>=1.15.2               # For database migrations
aiosqlite>=0.18.0
asyncpg>=0.27.0

# Flask-Migrate for database migrations
flask-migrate>=4.0.0

# Utilities
python-dotenv>=1.0.0
requests==2.31.0              # Already latest
aiohttp>=3.8.4
python-multipart>=0.0.6
apscheduler>=3.10.0           # For scheduling data collection tasks

# Email handling
aiosmtplib>=2.0.1

# Logging
colorlog>=6.7.0

# Password Hashing
passlib>=1.7.4  # For secure password hashing

# JWT Authentication
python-jose>=3.3.0  # For handling JSON Web Tokens
